requirements.txt
django >= 6.0.1
djangorestframework >= 3.16.1
python-dotenv
mysqlclient
pymysql
django-cors-headers
